refactor(#3271): split generators-native.ts god-file + DRY cleanup
Extract two cohesive subsystems out of the 4927-LOC generators-native.ts
(behaviour-preserving; emitted Wasm byte-for-byte identical across the
39 gc/standalone/wasi corpus emits, proven via prove-emit-identity):

- generators-native-consumer.ts (~1128 LOC): the native-generator CONSUMER /
  call-site subsystem — .next()/.return()/.throw() dispatch, {value,done}
  result-struct reads, and for-of / spread / toVec draining. Clean
  unidirectional cut (0 back-edges into the planner/state-machine core); the
  core re-exports its public entry points so external importers are unaffected.
- generators-native-ast-scan.ts (~315 LOC): the pure AST-scan predicate
  primitives (contains-yield/return, references this/super/own-name, try-region
  across yield, spill-safe ValType). ts + ValType + type-only GeneratorDecl deps.

DRY dedups (all provably byte-identical, spread-helper form):
- loadCastState (17 sites), readResultField (4 sites) in the consumer module;
- isFunctionLikeScope (7 sites) in the ast-scan module (the 6-way this/super
  variants that omit arrows are deliberately NOT folded).

tsc 0; smoke test tests/issue-3271.test.ts drives each relocated consumer path.

Refs #3271 #3182 ✓
